# LMSConnector/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: LMSConnector API
  version: 0.1.0
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /sources:
    get:
      summary: List LMS sources
      parameters:
        - in: header
          name: X-API-Key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of sources
    post:
      summary: Create LMS source
      parameters:
        - in: header
          name: X-API-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [moodle]
                base_url:
                  type: string
                token:
                  type: string
                active:
                  type: boolean
              required: [type, base_url]
      responses:
        '201':
          description: Created
  /sync/moodle:
    post:
      summary: Trigger Moodle sync for all active sources
      parameters:
        - in: header
          name: X-API-Key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sync results
  /sync/moodle/{sourceId}:
    post:
      summary: Trigger Moodle sync for a specific source
      parameters:
        - in: header
          name: X-API-Key
          required: true
          schema:
            type: string
        - in: path
          name: sourceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sync result
  /jobs/{jobId}:
    get:
      summary: Get job status
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status